install cmd line tools if they don't exist

+++ src/sdk-installer.ts
@@ -5,18 +5,37 @@ const BUILD_TOOLS_VERSION = '30.0.0';
 const CMDLINE_TOOLS_URL_MAC = 'https://dl.google.com/android/repository/commandlinetools-mac-6514223_latest.zip';
 const CMDLINE_TOOLS_URL_LINUX = 'https://dl.google.com/android/repository/commandlinetools-linux-6514223_latest.zip';
 
+async function hasCommandlineTools() {
+  let myOutput = '';
+  let myError = '';
+
+  const options = {
+    listeners : {
+      stdout: (data: Buffer) => {
+        myOutput += data.toString();
+      },
+      stderr: (data: Buffer) => {
+        myError += data.toString();
+      }
+    }
+  };
+  await exec.exec(`sudo ls ${process.env.ANDROID_HOME}/cmdline-tools`);
+  return myError != '';
+}
 /**
  * Installs & updates the Android SDK for the macOS platform, including SDK platform for the chosen API level, latest build tools, platform tools, Android Emulator,
  * and the system image for the chosen API level, CPU arch, and target.
  */
 export async function installAndroidSdk(apiLevel: number, target: string, arch: string, emulatorBuild?: string, ndkVersion?: string, cmakeVersion?: string): Promise<void> {
   const isOnMac = process.platform === 'darwin';
-  console.log('Installing new cmdline-tools.');
-  const sdkUrl = isOnMac ? CMDLINE_TOOLS_URL_MAC : CMDLINE_TOOLS_URL_LINUX;
-  await exec.exec(`sudo mkdir ${process.env.ANDROID_HOME}/cmdline-tools`);
-  await exec.exec(`curl -fo commandlinetools.zip ${sdkUrl}`);
-  await exec.exec(`sudo unzip -q commandlinetools.zip -d ${process.env.ANDROID_HOME}/cmdline-tools`);
-  await exec.exec(`sudo rm -f commandlinetools.zip`);
+  if (! (await hasCommandlineTools())) {
+    console.log('Installing new cmdline-tools.');
+    const sdkUrl = isOnMac ? CMDLINE_TOOLS_URL_MAC : CMDLINE_TOOLS_URL_LINUX;
+    await exec.exec(`sudo mkdir ${process.env.ANDROID_HOME}/cmdline-tools`);
+    await exec.exec(`curl -fo commandlinetools.zip ${sdkUrl}`);
+    await exec.exec(`sudo unzip -q commandlinetools.zip -d ${process.env.ANDROID_HOME}/cmdline-tools`);
+    await exec.exec(`sudo rm -f commandlinetools.zip`);
+  }
 
   // add paths for commandline-tools and platform-tools
   core.addPath(`${process.env.ANDROID_HOME}/cmdline-tools/tools:${process.env.ANDROID_HOME}/cmdline-tools/tools/bin:${process.env.ANDROID_HOME}/platform-tools`);
